Rework binding to the global environment (WIP)
